refactor: :recycle: rename "pages" folder to "features"
