+++ RespiraApp/app/index.tsx
@@ -1,4 +1,4 @@
-import { LoginPage } from "@/pages";
+import { LoginPage } from "@/features";
 
 export default function Index() {
   return <LoginPage />;
