+++ RespiraApp/app/tabs/profile/index.tsx
@@ -1,4 +1,4 @@
-import { ProfilePage } from "@/pages";
+import { ProfilePage } from "@/features";
 
 export default function Profile() {
   return <ProfilePage />;
