+++ RespiraApp/app/tabs/search/index.tsx
@@ -1,4 +1,4 @@
-import { SearchPage } from "@/pages";
+import { SearchPage } from "@/features";
 
 export default function Search() {
   return <SearchPage />;
